+++ examples/v2/next-pages-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkitnext/next-pages-router
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- Updated dependencies [d77f347]
+  - @copilotkitnext/react@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes