+++ packages/v1/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkit/a2ui-renderer
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- Updated dependencies [d77f347]
+  - @copilotkit/react-core@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes