+++ packages/v1/react-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ui
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- d77f347: Added in the useInterrupt hook
+- Updated dependencies [d77f347]
+  - @copilotkitnext/react@1.52.0-next.7
+  - @copilotkitnext/core@1.52.0-next.7
+  - @copilotkit/runtime-client-gql@1.52.0-next.7
+  - @copilotkit/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes