+++ packages/v1/react-textarea/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ui
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- Updated dependencies [d77f347]
+  - @copilotkit/react-core@1.52.0-next.7
+  - @copilotkit/runtime-client-gql@1.52.0-next.7
+  - @copilotkit/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes