+++ packages/v1/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkit/runtime-client-gql
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkit/runtime@1.52.0-next.7
+- @copilotkit/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes