+++ packages/v1/sdk-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/sdk-js
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes