+++ packages/v2/angular/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @copilotkitnext/angular
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- Updated dependencies [d77f347]
+  - @copilotkitnext/core@1.52.0-next.7
+  - @copilotkitnext/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes