+++ packages/v2/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkitnext/core
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- d77f347: Added in the useInterrupt hook
+  - @copilotkitnext/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes