+++ packages/v2/runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkitnext/runtime
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkitnext/shared@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes