+++ packages/v2/shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @copilotkitnext/shared
 
+## 1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ## 1.52.0-next.5