+++ packages/v2/voice/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/voice
 
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes